requirements.txt
fastapi>=0.103.0
uvicorn[standard]==0.22.0
pydantic>=2.0

numpy
joblib
scikit-learn==1.6.1
tensorflow-cpu==2.15.0

httpx
pytest